search,fetch functinality added for highvalue filters

+++ src/main/java/com/smsa/highValueAlerts/entity/SmsaRecepientMaster.java
@@ -26,7 +26,7 @@ public class SmsaRecepientMaster implements Serializable{
     private String smsaSenderBic;
 
     @Column(name = "SMSA_MSG_TYPE")
-    private Integer smsaMsgType;
+    private Long smsaMsgType;
 
     @Column(name = "SMSA_EMP_NAME", length = 150)
     private String smsaEmpName;
@@ -86,11 +86,11 @@ public class SmsaRecepientMaster implements Serializable{
         this.smsaSenderBic = smsaSenderBic;
     }
 
-    public Integer getSmsaMsgType() {
+    public Long getSmsaMsgType() {
         return smsaMsgType;
     }
 
-    public void setSmsaMsgType(Integer smsaMsgType) {
+    public void setSmsaMsgType(Long smsaMsgType) {
         this.smsaMsgType = smsaMsgType;
     }
